web/requirements.txt
Django==3.2.5
matplotlib==3.5.1
pandas==1.3.5
Pillow==9.2.0
PyMySQL==1.0.2
pytz==2022.1
seaborn==0.11.2
XlsxWriter==3.0.3
uwsgi==2.0.18
openpyxl==3.0.10